SPSS syntax added
I create this SPSS syntax to help me iterate through a list of variables in the dataset and perform analysis depending on if it is a categorical variable or numerical variable (user-define). The output is saved with the variable name being part of the file name. The output is saved in excel format, the number is prefixed to the filename so that the sequence in the window follows the sequence in the SPSS data window. SPSS Python extension is used. Python language is used